+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## current
+### Added
+- `Subcontext`: `alwaysRedraw` option to only redraw children if current time changes when false
 ## 0.1.0
 ### Added
 - Everything